ch2/ch2_api/requirements.txt
argon2-cffi==19.1.0
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2018.11.29
cffi==1.12.2
Click==7.0
colorama==0.4.1
coverage==4.5.3
cryptography==2.6.1
ecdsa==0.13
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
pipenv==2018.11.26
pluggy==0.9.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
pylint==2.2.2
pymongo==3.7.2
pytest==4.4.0
python-jose==3.0.1
rsa==4.0
six==1.12.0
virtualenv==16.2.0
virtualenv-clone==0.4.0
Werkzeug==0.15.2
wrapt==1.10.11
